+++ webapp/javascript/services/base.ts
@@ -45,16 +45,22 @@ export type RequestError =
   | RequestIncompleteError
   | ResponseOkNotInJSONFormat;
 
-function mountURL(req: RequestInfo): string {
+function join(base: string, path: string): string {
+  path = path.replace(/^\/+/, '');
+  base = base.replace(/\/+$/, '');
+  return `${base}/${path}`;
+}
+
+export function mountURL(req: RequestInfo): string {
   const baseName = basename();
 
   if (baseName) {
     if (typeof req === 'string') {
-      return new URL(`${baseName}/${req}`, window.location.href).href;
+      return new URL(join(baseName, req), window.location.href).href;
     }
 
     // req is an object
-    return new URL(`${baseName}/${req.url}`, window.location.href).href;
+    return new URL(join(baseName, req.url), window.location.href).href;
   }
 
   // no basename
@@ -112,8 +118,8 @@ export async function request(
       // Check if it's 401 unauthorized error
       if (response.status === 401) {
         // TODO: Introduce some kind of interceptor (?)
-        if (window && window.location) {
-          window.location.href = '/login';
+        if (!/\/(login|signup)$/.test(window?.location?.pathname)) {
+          window.location.href = mountURL('/login');
         }
       }