+++ webapp/javascript/services/users.ts
@@ -1,14 +1,13 @@
 import { Result } from '@webapp/util/fp';
 import {
   Users,
-  parse,
   type User,
   userModel,
   passwordEncode,
 } from '@webapp/models/users';
 import type { ZodError } from 'zod';
 import { modelToResult } from '@webapp/models/utils';
-import { request } from './base';
+import { request, parseResponse } from './base';
 import type { RequestError } from './base';
 
 export interface FetchUsersError {
@@ -21,7 +20,7 @@ export async function fetchUsers(): Promise<
   const response = await request('/api/users');
 
   if (response.isOk) {
-    return parse(response.value);
+    return parseResponse(response, userModel);
   }
 
   return Result.err<Users, RequestError>(response.error);
@@ -79,6 +78,45 @@ export async function loadCurrentUser(): Promise<
   }
 
   return Result.err<User, RequestError>(response.error);
+}
+
+export async function logIn({
+  username,
+  password,
+}: {
+  username: string;
+  password: string;
+}): Promise<Result<unknown, RequestError | ZodError>> {
+  const response = await request(`/login`, {
+    method: 'POST',
+    body: JSON.stringify({ username, password: passwordEncode(password) }),
+  });
+  if (response.isOk) {
+    return Result.ok(true);
+  }
+
+  return Result.err<boolean, RequestError>(response.error);
+}
+
+export async function signUp(data: {
+  username: string;
+  password: string;
+  fullName: string;
+  email: string;
+}): Promise<Result<unknown, RequestError | ZodError>> {
+  const response = await request(`/signup`, {
+    method: 'POST',
+    body: JSON.stringify({
+      ...data,
+      name: data.username,
+      password: passwordEncode(data.password),
+    }),
+  });
+  if (response.isOk) {
+    return Result.ok(true);
+  }
+
+  return Result.err<boolean, RequestError>(response.error);
 }
 
 export async function changeMyPassword(